requirements.txt
streamlit>=1.28.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.11.0
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
streamlit-webrtc>=0.47.1
aiortc>=1.5.0
av>=10.0.0
ffmpeg-python>=0.2.0